+++ packages/react-devtools-shared/src/devtools/views/Components/types.js
@@ -71,6 +71,9 @@ export type InspectedElement = {|
   // Can view component source location.
   canViewSource: boolean,
 
+  // Does the component have legacy context attached to it.
+  hasLegacyContext: boolean,
+
   // Inspectable properties.
   context: Object | null,
   hooks: Object | null,
@@ -80,7 +83,7 @@ export type InspectedElement = {|
   // List of owners
   owners: Array<Owner> | null,
 
-  // Location of component in source coude.
+  // Location of component in source code.
   source: Source | null,
 
   type: ElementType,