+++ packages/react-devtools/CHANGELOG.md
@@ -8,7 +8,9 @@
   
   <!-- Upcoming changes go here -->
 #### Bug fixes
+* Props/state editor supports adding new values and changing value types. ([hristo-kanchev](https://github.com/hristo-kanchev) in [#16700](https://github.com/facebook/react/pull/16700))
 * Profiler correctly saves/exports profiling data in Firefox now. ([hristo-kanchev](https://github.com/hristo-kanchev) in [#16612](https://github.com/facebook/react/pull/16612))
+* Class components now show "legacy context" header (rather than "context") for legacy API. ([hristo-kanchev](https://github.com/hristo-kanchev) in [#16617](https://github.com/facebook/react/pull/16617))
 </details>
 
 ## 4.0.6 (August 26, 2019)